requirements.txt
# Sage / pywaggle
pywaggle>=0.56.3

# Core CV
numpy>=1.24
Pillow>=10.0
opencv-python-headless>=4.8
requests>=2.31
huggingface_hub>=0.20

# YOLO + BioCLIP (torch pins: use matching wheels for your platform / CUDA)
torch
torchvision
ultralytics>=8.0
open_clip_torch>=2.24
